+++ packages/break-check/__tests__/fixtures/bun/private-method.test.js
